разбил написанный код на несколько модулей и доработал функцию по выдаче массива случайной длины, чтобы значения в массиве не повторялись
